Moving a2s.A2SEvent to java.awt
a2s.A2SEvent calls component.getMouseEventTarget, which is package-local
in Java 8.
