+++ sources/net.sf.j2s.java.core/src/a2s/A2SListener.java
@@ -1,5 +1,6 @@
 package a2s;
 
+import java.awt.A2SEvent;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.awt.event.AdjustmentEvent;
