+++ sources/net.sf.j2s.java.core/src/a2s/Checkbox.java
@@ -1,5 +1,6 @@
 package a2s;
 
+import java.awt.A2SEvent;
 import java.awt.HeadlessException;
 import java.awt.event.ActionEvent;
