+++ sources/net.sf.j2s.java.core/src/java/awt/A2SEvent.java
@@ -1,10 +1,5 @@
-package a2s;
+package java.awt;
 
-import java.awt.AWTEvent;
-import java.awt.Component;
-import java.awt.Container;
-import java.awt.Event;
-import java.awt.Point;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.awt.event.AdjustmentEvent;
@@ -22,6 +17,15 @@ import javax.swing.AbstractButton;
 import javax.swing.JComboBox;
 import javax.swing.JComponent;
 import javax.swing.JScrollBar;
+
+import a2s.A2SContainer;
+import a2s.A2SListener;
+import a2s.Choice;
+import a2s.Dialog;
+import a2s.Frame;
+import a2s.List;
+import a2s.MenuItem;
+import a2s.TextField;
 
 
 public class A2SEvent implements Runnable {
